docker vllm: Split entrypoints in separate clases and update vllm installation in docker (#1602)
docker vllm: Split entrypoints in separate clases and update vllm
installation in docker

---------
