+++ requirements/hpu.txt
@@ -8,3 +8,7 @@ triton==3.1.0
 setuptools>=77.0.3
 setuptools-scm>=8
 vllm-hpu-extension @ git+https://github.com/HabanaAI/vllm-hpu-extension.git@f831cb1
+
+# Dependencies for HPU vllm docker image
+datasets
+pandas
